+++ packages/components/src/Card/tests/Card.test.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import renderer from "react-test-renderer";
 import { fireEvent, render } from "@testing-library/react";
-import { Card } from ".";
+import { Card } from "..";
 
 it("renders a simple card", () => {
   const tree = renderer
